+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changes:
 * Better handling of accounts
 * Better icon for forked repo in event cell
 * Scrollable account form
+* Display merge button with status, even if the pull request is not mergeable
 
 Bugfixes: